fix(dev): split email-preview Client & force-dynamic; disable in CI
